fix move Cart and Checkout to Dashboard folder
