+++ client/src/routes/Main.tsx
@@ -13,8 +13,8 @@ import EnrollmentsLayout from '@/pages/Dashboard/Enrollments';
 import CourseLayout from '@/pages/Dashboard/CourseView';
 import CourseDetails from '../pages/CourseDetails';
 import CourseEditor from '../pages/Dashboard/CourseEditor';
-import Cart from '../pages/Cart';
-import Checkout from '../pages/Checkout';
+import Cart from '../pages/Dashboard/Cart';
+import Checkout from '../pages/Dashboard/Checkout';
 import UserProfile from '../pages/Dashboard/UserProfile';
 // import { Suspense } from 'react';
 // import { Loader } from 'lucide-react';
